Started work on a physical memory manager. Also we upgraded itoa to 64 bit and removed the intentional idt test kernel panic

+++ src/include/helpful.h
@@ -52,8 +52,8 @@ extern uint8_t *font_buffer;
 extern uint32_t text_color;
 
 void reverse(char s[]);
-void itoa_hex(uint32_t n, char s[]);
-void itoa(uint32_t n, char s[]);
+void itoa_hex(uint64_t n, char s[]);
+void itoa(uint64_t n, char s[]);
 size_t strlen(const char* str);
 
 void colon_three();
